Breaks things up, adds epub
