+++ requirements.txt
@@ -15,5 +15,8 @@ lxml
 # for rendering:
 WeasyPrint
 
+# for epub creation
+ebooklib
+
 # for sending to remarkable:
 git+https://github.com/j6k4m8/rmapy
