+++ setup.py
@@ -35,10 +35,9 @@ setup(
         "Programming Language :: Python :: 3.4",
         "Programming Language :: Python :: 3.5",
     ],
-    keywords=["sqlite"],
     packages=find_packages(exclude=["docs", "tests*"]),
     include_package_data=True,
-    entry_points={"console_scripts": ["goosepaper=goosepaper:cli", "upload_to_remarkable=goosepaper:upload"],},
+    entry_points={"console_scripts": ["goosepaper=goosepaper.__main__:main", "upload_to_remarkable=goosepaper.upload:upload"],},
     author="Jordan Matelsky",
     install_requires=install_requires,
     dependency_links=dependency_links,
